+++ platform/core.windows/src/org/netbeans/core/windows/view/SplitView.java
@@ -151,7 +151,7 @@ public class SplitView extends ViewElement {
             }
             Integer override = Integer.getInteger("Nb.SplitPane.dividerSize"); //NOI18N
             if( null != override ) {
-                dividerSize = override.intValue();
+                dividerSize = override;
             }
             splitPane.setDividerSize(dividerSize);
